baseimgs/binary/requirements.txt
pyinstaller
GitPython
matplotlib